Move Spec<->API converter functions to API files (#3415)
Move Spec<->API converter functions to API files
