+++ src/deploy/functions/backend.ts
@@ -274,299 +274,6 @@ export function scheduleIdForFunction(cloudFunction: TargetIds) {
   return `firebase-schedule-${cloudFunction.id}-${cloudFunction.region}`;
 }
 
-/**
- * Convert the API agnostic FunctionSpec struct to a CloudFunction proto for the v1 API.
- */
-export function toGCFv1Function(
-  cloudFunction: FunctionSpec,
-  sourceUploadUrl: string
-): Omit<gcf.CloudFunction, gcf.OutputOnlyFields> {
-  if (cloudFunction.apiVersion != 1) {
-    throw new FirebaseError(
-      "Trying to create a v1 CloudFunction with v2 API. This should never happen"
-    );
-  }
-
-  if (!isValidRuntime(cloudFunction.runtime)) {
-    throw new FirebaseError(
-      "Failed internal assertion. Trying to deploy a new function with a deprecated runtime." +
-        " This should never happen"
-    );
-  }
-  const gcfFunction: Omit<gcf.CloudFunction, gcf.OutputOnlyFields> = {
-    name: functionName(cloudFunction),
-    sourceUploadUrl: sourceUploadUrl,
-    entryPoint: cloudFunction.entryPoint,
-    runtime: cloudFunction.runtime,
-  };
-
-  if (isEventTrigger(cloudFunction.trigger)) {
-    gcfFunction.eventTrigger = {
-      eventType: cloudFunction.trigger.eventType,
-      resource: cloudFunction.trigger.eventFilters.resource,
-      // Service is unnecessary and deprecated
-    };
-
-    // For field masks to pick up a deleted failure policy we must inject an undefined
-    // when retry is false
-    gcfFunction.eventTrigger.failurePolicy = cloudFunction.trigger.retry
-      ? { retry: {} }
-      : undefined;
-  } else {
-    gcfFunction.httpsTrigger = {
-      securityLevel: cloudFunction.trigger.allowInsecure ? "SECURE_OPTIONAL" : "SECURE_ALWAYS",
-    };
-  }
-
-  proto.copyIfPresent(
-    gcfFunction,
-    cloudFunction,
-    "serviceAccountEmail",
-    "timeout",
-    "availableMemoryMb",
-    "minInstances",
-    "maxInstances",
-    "vpcConnector",
-    "vpcConnectorEgressSettings",
-    "ingressSettings",
-    "labels",
-    "environmentVariables"
-  );
-
-  return gcfFunction;
-}
-
-/**
- * Converts a Cloud Function from the v1 API into a version-agnostic FunctionSpec struct.
- * This API exists outside the GCF namespace because GCF returns an Operation<CloudFunction>
- * and code may have to call this method explicitly.
- */
-export function fromGCFv1Function(gcfFunction: gcf.CloudFunction): FunctionSpec {
-  const [, project, , region, , id] = gcfFunction.name.split("/");
-  let trigger: EventTrigger | HttpsTrigger;
-  let uri: string | undefined;
-  if (gcfFunction.httpsTrigger) {
-    trigger = {
-      // Note: default (empty) value intentionally means true
-      allowInsecure: gcfFunction.httpsTrigger.securityLevel !== "SECURE_ALWAYS",
-    };
-    uri = gcfFunction.httpsTrigger.url;
-  } else {
-    trigger = {
-      eventType: gcfFunction.eventTrigger!.eventType,
-      eventFilters: {
-        resource: gcfFunction.eventTrigger!.resource,
-      },
-      retry: !!gcfFunction.eventTrigger!.failurePolicy?.retry,
-    };
-  }
-
-  if (!isValidRuntime(gcfFunction.runtime)) {
-    logger.debug("GCFv1 function has a deprecated runtime:", JSON.stringify(gcfFunction, null, 2));
-  }
-
-  const cloudFunction: FunctionSpec = {
-    apiVersion: 1,
-    id,
-    project,
-    region,
-    trigger,
-    entryPoint: gcfFunction.entryPoint,
-    runtime: gcfFunction.runtime,
-  };
-  if (uri) {
-    cloudFunction.uri = uri;
-  }
-  proto.copyIfPresent(
-    cloudFunction,
-    gcfFunction,
-    "serviceAccountEmail",
-    "availableMemoryMb",
-    "timeout",
-    "minInstances",
-    "maxInstances",
-    "vpcConnector",
-    "vpcConnectorEgressSettings",
-    "ingressSettings",
-    "labels",
-    "environmentVariables",
-    "sourceUploadUrl"
-  );
-
-  return cloudFunction;
-}
-
-export function toGCFv2Function(cloudFunction: FunctionSpec, source: gcfV2.StorageSource) {
-  if (cloudFunction.apiVersion != 2) {
-    throw new FirebaseError(
-      "Trying to create a v2 CloudFunction with v1 API. This should never happen"
-    );
-  }
-
-  if (!isValidRuntime(cloudFunction.runtime)) {
-    throw new FirebaseError(
-      "Failed internal assertion. Trying to deploy a new function with a deprecated runtime." +
-        " This should never happen"
-    );
-  }
-
-  const gcfFunction: Omit<gcfV2.CloudFunction, gcfV2.OutputOnlyFields> = {
-    name: functionName(cloudFunction),
-    buildConfig: {
-      runtime: cloudFunction.runtime,
-      entryPoint: cloudFunction.entryPoint,
-      source: {
-        storageSource: source,
-      },
-      // We don't use build environment variables,
-      environmentVariables: {},
-    },
-    serviceConfig: {},
-  };
-
-  proto.copyIfPresent(
-    gcfFunction.serviceConfig,
-    cloudFunction,
-    "availableMemoryMb",
-    "environmentVariables",
-    "vpcConnector",
-    "vpcConnectorEgressSettings",
-    "serviceAccountEmail",
-    "ingressSettings"
-  );
-  proto.renameIfPresent(
-    gcfFunction.serviceConfig,
-    cloudFunction,
-    "timeoutSeconds",
-    "timeout",
-    proto.secondsFromDuration
-  );
-  proto.renameIfPresent(
-    gcfFunction.serviceConfig,
-    cloudFunction,
-    "minInstanceCount",
-    "minInstances"
-  );
-  proto.renameIfPresent(
-    gcfFunction.serviceConfig,
-    cloudFunction,
-    "maxInstanceCount",
-    "maxInstances"
-  );
-
-  if (isEventTrigger(cloudFunction.trigger)) {
-    gcfFunction.eventTrigger = {
-      eventType: cloudFunction.trigger.eventType,
-    };
-    if (gcfFunction.eventTrigger.eventType === gcfV2.PUBSUB_PUBLISH_EVENT) {
-      gcfFunction.eventTrigger.pubsubTopic = cloudFunction.trigger.eventFilters.resource;
-    } else {
-      gcfFunction.eventTrigger.eventFilters = [];
-      for (const [attribute, value] of Object.entries(cloudFunction.trigger.eventFilters)) {
-        gcfFunction.eventTrigger.eventFilters.push({ attribute, value });
-      }
-    }
-
-    if (cloudFunction.trigger.retry) {
-      logger.warn("Cannot set a retry policy on Cloud Function", cloudFunction.id);
-    }
-  } else if (cloudFunction.trigger.allowInsecure) {
-    logger.warn("Cannot enable insecure traffic for Cloud Function", cloudFunction.id);
-  }
-  proto.copyIfPresent(gcfFunction, cloudFunction, "labels");
-
-  return gcfFunction;
-}
-
-export function fromGCFv2Function(gcfFunction: gcfV2.CloudFunction): FunctionSpec {
-  const [, project, , region, , id] = gcfFunction.name.split("/");
-  let trigger: EventTrigger | HttpsTrigger;
-  if (gcfFunction.eventTrigger) {
-    trigger = {
-      eventType: gcfFunction.eventTrigger!.eventType,
-      eventFilters: {},
-      retry: false,
-    };
-    if (gcfFunction.eventTrigger.pubsubTopic) {
-      trigger.eventFilters.resource = gcfFunction.eventTrigger.pubsubTopic;
-    } else {
-      for (const { attribute, value } of gcfFunction.eventTrigger.eventFilters || []) {
-        trigger.eventFilters[attribute] = value;
-      }
-    }
-  } else {
-    trigger = {
-      allowInsecure: false,
-    };
-  }
-
-  if (!isValidRuntime(gcfFunction.buildConfig.runtime)) {
-    logger.debug("GCFv2 function has a deprecated runtime:", JSON.stringify(gcfFunction, null, 2));
-  }
-
-  const cloudFunction: FunctionSpec = {
-    apiVersion: 2,
-    id,
-    project,
-    region,
-    trigger,
-    entryPoint: gcfFunction.buildConfig.entryPoint,
-    runtime: gcfFunction.buildConfig.runtime,
-    uri: gcfFunction.serviceConfig.uri,
-  };
-  proto.copyIfPresent(
-    cloudFunction,
-    gcfFunction.serviceConfig,
-    "serviceAccountEmail",
-    "availableMemoryMb",
-    "vpcConnector",
-    "vpcConnectorEgressSettings",
-    "ingressSettings",
-    "environmentVariables"
-  );
-  proto.renameIfPresent(
-    cloudFunction,
-    gcfFunction.serviceConfig,
-    "timeout",
-    "timeoutSeconds",
-    proto.durationFromSeconds
-  );
-  proto.renameIfPresent(
-    cloudFunction,
-    gcfFunction.serviceConfig,
-    "minInstances",
-    "minInstanceCount"
-  );
-  proto.renameIfPresent(
-    cloudFunction,
-    gcfFunction.serviceConfig,
-    "maxInstances",
-    "maxInstanceCount"
-  );
-  proto.copyIfPresent(cloudFunction, gcfFunction, "labels");
-
-  return cloudFunction;
-}
-
-/** Converts a version agnostic ScheduleSpec to a CloudScheduler v1 Job. */
-export function toJob(schedule: ScheduleSpec, appEngineLocation: string): cloudscheduler.Job {
-  const job: cloudscheduler.Job = {
-    name: scheduleName(schedule, appEngineLocation),
-    schedule: schedule.schedule!,
-  };
-  proto.copyIfPresent(job, schedule, "timeZone", "retryConfig");
-  if (schedule.transport === "https") {
-    throw new FirebaseError("HTTPS transport for scheduled functions is not yet supported");
-  }
-  job.pubsubTarget = {
-    topicName: topicName(schedule),
-    attributes: {
-      scheduled: "true",
-    },
-  };
-  return job;
-}
-
 interface PrivateContextFields {
   existingBackend: Backend;
   loadedExistingBackend?: boolean;
@@ -616,7 +323,7 @@ async function loadExistingBackend(ctx: Context & PrivateContextFields): Promise
   };
   const gcfV1Results = await gcf.listAllFunctions(ctx.projectId);
   for (const apiFunction of gcfV1Results.functions) {
-    const specFunction = fromGCFv1Function(apiFunction);
+    const specFunction = gcf.specFromFunction(apiFunction);
     ctx.existingBackend.cloudFunctions.push(specFunction);
     const isScheduled = apiFunction.labels?.["deployment-scheduled"] === "true";
     if (isScheduled) {
@@ -651,7 +358,7 @@ async function loadExistingBackend(ctx: Context & PrivateContextFields): Promise
 
   const gcfV2Results = await gcfV2.listAllFunctions(ctx.projectId);
   for (const apiFunction of gcfV2Results.functions) {
-    const specFunction = fromGCFv2Function(apiFunction);
+    const specFunction = gcfV2.specFromFunction(apiFunction);
     ctx.existingBackend.cloudFunctions.push(specFunction);
     const pubsubScheduled = apiFunction.labels?.["deployment-scheduled"] === "true";
     const httpsScheduled = apiFunction.labels?.["deployment-scheduled"] === "https";
