+++ src/test/deploy/functions/backend.spec.ts
@@ -8,7 +8,6 @@ import * as backend from "../../../deploy/functions/backend";
 import * as gcf from "../../../gcp/cloudfunctions";
 import * as gcfV2 from "../../../gcp/cloudfunctionsv2";
 import * as utils from "../../../utils";
-import { Context } from "mocha";
 
 describe("Backend", () => {
   const FUNCTION_NAME: backend.TargetIds = {
@@ -145,528 +144,6 @@ describe("Backend", () => {
       expect(matcher({ ...FUNCTION_SPEC, id: "other" })).to.be.false;
       expect(matcher({ ...FUNCTION_SPEC, region: "other" })).to.be.false;
       expect(matcher({ ...FUNCTION_SPEC, project: "other" })).to.be.false;
-    });
-  });
-
-  describe("toGCFv1Function", () => {
-    const UPLOAD_URL = "https://storage.googleapis.com/projects/-/buckets/sample/source.zip";
-    it("should guard against version mixing", () => {
-      expect(() => {
-        backend.toGCFv1Function({ ...FUNCTION_SPEC, apiVersion: 2 }, UPLOAD_URL);
-      }).to.throw;
-    });
-
-    it("should copy a minimal function", () => {
-      expect(backend.toGCFv1Function(FUNCTION_SPEC, UPLOAD_URL)).to.deep.equal({
-        ...CLOUD_FUNCTION,
-        sourceUploadUrl: UPLOAD_URL,
-        httpsTrigger: {
-          securityLevel: "SECURE_ALWAYS",
-        },
-      });
-
-      const eventFunction = {
-        ...FUNCTION_SPEC,
-        trigger: {
-          eventType: "google.pubsub.topic.publish",
-          eventFilters: {
-            resource: "projects/p/topics/t",
-          },
-          retry: false,
-        },
-      };
-      const eventGcfFunction = {
-        ...CLOUD_FUNCTION,
-        sourceUploadUrl: UPLOAD_URL,
-        eventTrigger: {
-          eventType: "google.pubsub.topic.publish",
-          resource: "projects/p/topics/t",
-          failurePolicy: undefined,
-        },
-      };
-      expect(backend.toGCFv1Function(eventFunction, UPLOAD_URL)).to.deep.equal(eventGcfFunction);
-    });
-
-    it("should copy trival fields", () => {
-      const fullFunction: backend.FunctionSpec = {
-        ...FUNCTION_SPEC,
-        availableMemoryMb: 128,
-        minInstances: 1,
-        maxInstances: 42,
-        vpcConnector: "connector",
-        vpcConnectorEgressSettings: "ALL_TRAFFIC",
-        ingressSettings: "ALLOW_ALL",
-        timeout: "15s",
-        serviceAccountEmail: "inlined@google.com",
-        labels: {
-          foo: "bar",
-        },
-        environmentVariables: {
-          FOO: "bar",
-        },
-      };
-
-      const fullGcfFunction: Omit<gcf.CloudFunction, gcf.OutputOnlyFields> = {
-        ...CLOUD_FUNCTION,
-        sourceUploadUrl: UPLOAD_URL,
-        httpsTrigger: {
-          securityLevel: "SECURE_ALWAYS",
-        },
-        labels: {
-          foo: "bar",
-        },
-        environmentVariables: {
-          FOO: "bar",
-        },
-        maxInstances: 42,
-        minInstances: 1,
-        vpcConnector: "connector",
-        vpcConnectorEgressSettings: "ALL_TRAFFIC",
-        ingressSettings: "ALLOW_ALL",
-        availableMemoryMb: 128,
-        timeout: "15s",
-        serviceAccountEmail: "inlined@google.com",
-      };
-
-      expect(backend.toGCFv1Function(fullFunction, UPLOAD_URL)).to.deep.equal(fullGcfFunction);
-    });
-
-    it("should calculate non-trivial fields", () => {
-      const complexFunction: backend.FunctionSpec = {
-        ...FUNCTION_SPEC,
-        trigger: {
-          eventType: "google.pubsub.topic.publish",
-          eventFilters: {
-            resource: "projects/p/topics/t",
-          },
-          retry: true,
-        },
-      };
-
-      const complexGcfFunction: Omit<gcf.CloudFunction, gcf.OutputOnlyFields> = {
-        ...CLOUD_FUNCTION,
-        sourceUploadUrl: UPLOAD_URL,
-        eventTrigger: {
-          eventType: "google.pubsub.topic.publish",
-          resource: "projects/p/topics/t",
-          failurePolicy: {
-            retry: {},
-          },
-        },
-      };
-
-      expect(backend.toGCFv1Function(complexFunction, UPLOAD_URL)).to.deep.equal(
-        complexGcfFunction
-      );
-    });
-  });
-
-  describe("fromGCFv1Function", () => {
-    it("should copy a minimal version", () => {
-      expect(
-        backend.fromGCFv1Function({
-          ...HAVE_CLOUD_FUNCTION,
-          httpsTrigger: {
-            securityLevel: "SECURE_ALWAYS",
-          },
-        })
-      ).to.deep.equal(FUNCTION_SPEC);
-    });
-
-    it("should translate event triggers", () => {
-      expect(
-        backend.fromGCFv1Function({
-          ...HAVE_CLOUD_FUNCTION,
-          eventTrigger: {
-            eventType: "google.pubsub.topic.publish",
-            resource: "projects/p/topics/t",
-            failurePolicy: {
-              retry: {},
-            },
-          },
-        })
-      ).to.deep.equal({
-        ...FUNCTION_SPEC,
-        trigger: {
-          eventType: "google.pubsub.topic.publish",
-          eventFilters: {
-            resource: "projects/p/topics/t",
-          },
-          retry: true,
-        },
-      });
-
-      // And again w/o the failure policy
-      expect(
-        backend.fromGCFv1Function({
-          ...HAVE_CLOUD_FUNCTION,
-          eventTrigger: {
-            eventType: "google.pubsub.topic.publish",
-            resource: "projects/p/topics/t",
-          },
-        })
-      ).to.deep.equal({
-        ...FUNCTION_SPEC,
-        trigger: {
-          eventType: "google.pubsub.topic.publish",
-          eventFilters: {
-            resource: "projects/p/topics/t",
-          },
-          retry: false,
-        },
-      });
-    });
-
-    it("should copy optional fields", () => {
-      const extraFields: Partial<backend.FunctionSpec> = {
-        availableMemoryMb: 128,
-        minInstances: 1,
-        maxInstances: 42,
-        vpcConnector: "connector",
-        vpcConnectorEgressSettings: "ALL_TRAFFIC",
-        ingressSettings: "ALLOW_ALL",
-        serviceAccountEmail: "inlined@google.com",
-        timeout: "15s",
-        labels: {
-          foo: "bar",
-        },
-        environmentVariables: {
-          FOO: "bar",
-        },
-      };
-      expect(
-        backend.fromGCFv1Function({
-          ...HAVE_CLOUD_FUNCTION,
-          ...extraFields,
-          httpsTrigger: {},
-        } as gcf.CloudFunction)
-      ).to.deep.equal({
-        ...FUNCTION_SPEC,
-        ...extraFields,
-        trigger: {
-          allowInsecure: true,
-        },
-      });
-    });
-
-    it("should transform fields", () => {
-      expect(
-        backend.fromGCFv1Function({
-          ...HAVE_CLOUD_FUNCTION,
-          httpsTrigger: {
-            securityLevel: "SECURE_OPTIONAL",
-          },
-        })
-      ).to.deep.equal({
-        ...FUNCTION_SPEC,
-        trigger: {
-          allowInsecure: true,
-        },
-      });
-    });
-  });
-
-  describe("toGCFv2Function", () => {
-    const UPLOAD_URL = "https://storage.googleapis.com/projects/-/buckets/sample/source.zip";
-    it("should guard against version mixing", () => {
-      expect(() => {
-        backend.toGCFv2Function({ ...FUNCTION_SPEC, apiVersion: 1 }, CLOUD_FUNCTION_V2_SOURCE);
-      }).to.throw;
-    });
-
-    it("should copy a minimal function", () => {
-      expect(
-        backend.toGCFv2Function(
-          {
-            ...FUNCTION_SPEC,
-            apiVersion: 2,
-          },
-          CLOUD_FUNCTION_V2_SOURCE
-        )
-      ).to.deep.equal(CLOUD_FUNCTION_V2);
-
-      const eventFunction: backend.FunctionSpec = {
-        ...FUNCTION_SPEC,
-        apiVersion: 2,
-        trigger: {
-          eventType: "google.cloud.audit.log.v1.written",
-          eventFilters: {
-            resource: "projects/p/regions/r/instances/i",
-            serviceName: "compute.googleapis.com",
-          },
-          retry: false,
-        },
-      };
-      const eventGcfFunction: Omit<gcfV2.CloudFunction, gcfV2.OutputOnlyFields> = {
-        ...CLOUD_FUNCTION_V2,
-        eventTrigger: {
-          eventType: "google.cloud.audit.log.v1.written",
-          eventFilters: [
-            {
-              attribute: "resource",
-              value: "projects/p/regions/r/instances/i",
-            },
-            {
-              attribute: "serviceName",
-              value: "compute.googleapis.com",
-            },
-          ],
-        },
-      };
-      expect(backend.toGCFv2Function(eventFunction, CLOUD_FUNCTION_V2_SOURCE)).to.deep.equal(
-        eventGcfFunction
-      );
-    });
-
-    it("should copy trival fields", () => {
-      const fullFunction: backend.FunctionSpec = {
-        ...FUNCTION_SPEC,
-        apiVersion: 2,
-        availableMemoryMb: 128,
-        vpcConnector: "connector",
-        vpcConnectorEgressSettings: "ALL_TRAFFIC",
-        ingressSettings: "ALLOW_ALL",
-        serviceAccountEmail: "inlined@google.com",
-        labels: {
-          foo: "bar",
-        },
-        environmentVariables: {
-          FOO: "bar",
-        },
-      };
-
-      const fullGcfFunction: Omit<gcfV2.CloudFunction, gcfV2.OutputOnlyFields> = {
-        ...CLOUD_FUNCTION_V2,
-        labels: {
-          foo: "bar",
-        },
-        serviceConfig: {
-          ...CLOUD_FUNCTION_V2.serviceConfig,
-          environmentVariables: {
-            FOO: "bar",
-          },
-          vpcConnector: "connector",
-          vpcConnectorEgressSettings: "ALL_TRAFFIC",
-          ingressSettings: "ALLOW_ALL",
-          availableMemoryMb: 128,
-          serviceAccountEmail: "inlined@google.com",
-        },
-      };
-
-      expect(backend.toGCFv2Function(fullFunction, CLOUD_FUNCTION_V2_SOURCE)).to.deep.equal(
-        fullGcfFunction
-      );
-    });
-
-    it("should calculate non-trivial fields", () => {
-      const complexFunction: backend.FunctionSpec = {
-        ...FUNCTION_SPEC,
-        apiVersion: 2,
-        trigger: {
-          eventType: gcfV2.PUBSUB_PUBLISH_EVENT,
-          eventFilters: {
-            resource: "projects/p/topics/t",
-          },
-          retry: false,
-        },
-        maxInstances: 42,
-        minInstances: 1,
-        timeout: "15s",
-      };
-
-      const complexGcfFunction: Omit<gcfV2.CloudFunction, gcfV2.OutputOnlyFields> = {
-        ...CLOUD_FUNCTION_V2,
-        eventTrigger: {
-          eventType: gcfV2.PUBSUB_PUBLISH_EVENT,
-          pubsubTopic: "projects/p/topics/t",
-        },
-        serviceConfig: {
-          ...CLOUD_FUNCTION_V2.serviceConfig,
-          maxInstanceCount: 42,
-          minInstanceCount: 1,
-          timeoutSeconds: 15,
-        },
-      };
-
-      expect(backend.toGCFv2Function(complexFunction, CLOUD_FUNCTION_V2_SOURCE)).to.deep.equal(
-        complexGcfFunction
-      );
-    });
-  });
-
-  describe("fromGCFv2Function", () => {
-    it("should copy a minimal version", () => {
-      expect(backend.fromGCFv2Function(HAVE_CLOUD_FUNCTION_V2)).to.deep.equal({
-        ...FUNCTION_SPEC,
-        apiVersion: 2,
-        uri: RUN_URI,
-      });
-    });
-
-    it("should translate event triggers", () => {
-      expect(
-        backend.fromGCFv2Function({
-          ...HAVE_CLOUD_FUNCTION_V2,
-          eventTrigger: {
-            eventType: gcfV2.PUBSUB_PUBLISH_EVENT,
-            pubsubTopic: "projects/p/topics/t",
-          },
-        })
-      ).to.deep.equal({
-        ...FUNCTION_SPEC,
-        apiVersion: 2,
-        uri: RUN_URI,
-        trigger: {
-          eventType: gcfV2.PUBSUB_PUBLISH_EVENT,
-          eventFilters: {
-            resource: "projects/p/topics/t",
-          },
-          retry: false,
-        },
-      });
-
-      // And again w/ a normal event trigger
-      expect(
-        backend.fromGCFv2Function({
-          ...HAVE_CLOUD_FUNCTION_V2,
-          eventTrigger: {
-            eventType: "google.cloud.audit.log.v1.written",
-            eventFilters: [
-              {
-                attribute: "resource",
-                value: "projects/p/regions/r/instances/i",
-              },
-              {
-                attribute: "serviceName",
-                value: "compute.googleapis.com",
-              },
-            ],
-          },
-        })
-      ).to.deep.equal({
-        ...FUNCTION_SPEC,
-        apiVersion: 2,
-        uri: RUN_URI,
-        trigger: {
-          eventType: "google.cloud.audit.log.v1.written",
-          eventFilters: {
-            resource: "projects/p/regions/r/instances/i",
-            serviceName: "compute.googleapis.com",
-          },
-          retry: false,
-        },
-      });
-    });
-
-    it("should copy optional fields", () => {
-      const extraFields: Partial<backend.FunctionSpec> = {
-        availableMemoryMb: 128,
-        vpcConnector: "connector",
-        vpcConnectorEgressSettings: "ALL_TRAFFIC",
-        ingressSettings: "ALLOW_ALL",
-        serviceAccountEmail: "inlined@google.com",
-        environmentVariables: {
-          FOO: "bar",
-        },
-      };
-      expect(
-        backend.fromGCFv2Function({
-          ...HAVE_CLOUD_FUNCTION_V2,
-          serviceConfig: {
-            ...HAVE_CLOUD_FUNCTION_V2.serviceConfig,
-            ...extraFields,
-          },
-          labels: {
-            foo: "bar",
-          },
-        })
-      ).to.deep.equal({
-        ...FUNCTION_SPEC,
-        apiVersion: 2,
-        uri: RUN_URI,
-        ...extraFields,
-        labels: {
-          foo: "bar",
-        },
-      });
-    });
-
-    it("should transform fields", () => {
-      const extraFields: Partial<backend.FunctionSpec> = {
-        minInstances: 1,
-        maxInstances: 42,
-        timeout: "15s",
-      };
-
-      const extraGcfFields: Partial<gcfV2.ServiceConfig> = {
-        minInstanceCount: 1,
-        maxInstanceCount: 42,
-        timeoutSeconds: 15,
-      };
-
-      expect(
-        backend.fromGCFv2Function({
-          ...HAVE_CLOUD_FUNCTION_V2,
-          serviceConfig: {
-            ...HAVE_CLOUD_FUNCTION_V2.serviceConfig,
-            ...extraGcfFields,
-          },
-        })
-      ).to.deep.equal({
-        ...FUNCTION_SPEC,
-        apiVersion: 2,
-        uri: RUN_URI,
-        ...extraFields,
-      });
-    });
-  });
-
-  describe("toJob", () => {
-    it("should copy minimal fields", () => {
-      expect(backend.toJob(SCHEDULE, "appEngineLocation")).to.deep.equal({
-        name: "projects/project/locations/appEngineLocation/jobs/firebase-schedule-id-region",
-        schedule: "every 1 minutes",
-        pubsubTarget: {
-          topicName: "projects/project/topics/firebase-schedule-id-region",
-          attributes: {
-            scheduled: "true",
-          },
-        },
-      });
-    });
-
-    it("should copy optional fields", () => {
-      expect(
-        backend.toJob(
-          {
-            ...SCHEDULE,
-            timeZone: "America/Los_Angeles",
-            retryConfig: {
-              maxDoublings: 2,
-              maxBackoffDuration: "20s",
-              minBackoffDuration: "1s",
-              maxRetryDuration: "60s",
-            },
-          },
-          "appEngineLocation"
-        )
-      ).to.deep.equal({
-        name: "projects/project/locations/appEngineLocation/jobs/firebase-schedule-id-region",
-        schedule: "every 1 minutes",
-        timeZone: "America/Los_Angeles",
-        retryConfig: {
-          maxDoublings: 2,
-          maxBackoffDuration: "20s",
-          minBackoffDuration: "1s",
-          maxRetryDuration: "60s",
-        },
-        pubsubTarget: {
-          topicName: "projects/project/topics/firebase-schedule-id-region",
-          attributes: {
-            scheduled: "true",
-          },
-        },
-      });
     });
   });
