+++ src/test/gcp/cloudscheduler.spec.ts
@@ -118,4 +118,64 @@ describe("cloudscheduler", () => {
       expect(nock.isDone()).to.be.true;
     });
   });
+
+  describe("toJob", () => {
+    const SCHEDULE = {
+      id: "firebase-schedule-id-region",
+      project: "project",
+      schedule: "every 1 minutes",
+      transport: "pubsub" as any,
+      targetService: {
+        id: "id",
+        region: "region",
+        project: "project",
+      },
+    };
+    it("should copy minimal fields", () => {
+      expect(cloudscheduler.jobFromSpec(SCHEDULE, "appEngineLocation")).to.deep.equal({
+        name: "projects/project/locations/appEngineLocation/jobs/firebase-schedule-id-region",
+        schedule: "every 1 minutes",
+        pubsubTarget: {
+          topicName: "projects/project/topics/firebase-schedule-id-region",
+          attributes: {
+            scheduled: "true",
+          },
+        },
+      });
+    });
+
+    it("should copy optional fields", () => {
+      expect(
+        cloudscheduler.jobFromSpec(
+          {
+            ...SCHEDULE,
+            timeZone: "America/Los_Angeles",
+            retryConfig: {
+              maxDoublings: 2,
+              maxBackoffDuration: "20s",
+              minBackoffDuration: "1s",
+              maxRetryDuration: "60s",
+            },
+          },
+          "appEngineLocation"
+        )
+      ).to.deep.equal({
+        name: "projects/project/locations/appEngineLocation/jobs/firebase-schedule-id-region",
+        schedule: "every 1 minutes",
+        timeZone: "America/Los_Angeles",
+        retryConfig: {
+          maxDoublings: 2,
+          maxBackoffDuration: "20s",
+          minBackoffDuration: "1s",
+          maxRetryDuration: "60s",
+        },
+        pubsubTarget: {
+          topicName: "projects/project/topics/firebase-schedule-id-region",
+          attributes: {
+            scheduled: "true",
+          },
+        },
+      });
+    });
+  });
 });
